chore(rust): 清理 77 条 clippy 告警（cargo clippy -- -D warnings 清零）
机械项用 cargo clippy --fix 批量修：map_or→is_some_and / from_ref 取代 &[x.clone()] /

format!(纯字面量)→to_string / 手写 impl Default→derive / match 单模式→if let / last()→next_back() 等。

设计类手工处理：too_many_arguments（8 处）加带说明的 #[allow]；type_complexity 抽 type 别名

(agent/mod.rs PersistSnapshotFn / BoxedPersistSnapshotFn)；large_enum_variant 给 ProviderBridgeMsg::Done.result 装箱；

doc 缩进类补空行断开 lazy continuation、模块头 /** */ → /*! */。

⚠️ clippy --fix 曾把 #[cfg(test)] 挪给新插入的 impl Default，导致 impl TestEventSink 在非 test 构建下 E0425，已改用 derive(Default) 修正。

验证：cargo clippy --workspace --all-targets -- -D warnings 0 告警；cargo test --workspace 582 passed。

+++ src-tauri/src/tray/mod.rs
@@ -903,7 +903,7 @@ mod tests {
             ]
         };
         // 红点圆心：右下角留 1px 边距、半径 = min(w,h)/4
-        let r = (32 / 4) as i32;
+        let r = 32 / 4;
         let c = 32 - r - 1;
         assert_eq!(px(c as u32, c as u32), [0xF2, 0x4D, 0x4D, 0xFF]);
         // 左上角保持原样
